Added a maintenance script
Added a maintenance script which deletes cached RFC documents which were
not accessed longer than a holdtime. The default is 30 days but can be
overriden to an argument to the script.
The script can also be used by user to maintain their own cache dir by
supplying the script with appropriate arguments.