Allow configure run for "debian/control"
Formerly it lead to an error state to get "debian/control" generated out
of a "debian/control.in" file because the normal behaving clean target
in the makefile deleted it in the normal debian package creation process
despite it was needed in the build process prelude.
Changed this so it works now.