Bypassing the use of m4 for help screen generation
Since we include the help screen as comment lines inside the script we
have to parse it to get the correct passage and to remove leading
comment signs. Formerly we used "sed" /and/ "m4" - the later to
substitute placeholders.
No we use "sed" alone since it certainly is also capable to replace
placeholders.