Corrected minor error concerning "dh_auto_clean"
When building a debian package out of the sources the build stack also
executes a "dh_auto_clean" which in turn also executes the target
"distclean" if it is present in the makefile. This leeds to the deletion
of the "debian/control" file which is needed /after/ the build process
=> Error.
Renamed this target so that the build process succeeds but a total
cleanup can still be initiated via a suitable call of the make file.