Cache maintenance is now done via tmpfiles.d
Systemd offers the systemd-tmpfiles mechanism which on one hand enshures
the existence of for instance a cache directory and on the other hand
removes files according to certain criteria like last modification or
access time.
Therewith we no longer need the cron.d entry which calls the explicit
maintenance script, we possibly no longer need this script - have to see
how to maintain user private caches - and we don't need the debian rule
override which enshures untypical access rights on the cache directory.