Package freshed up
Since the last bigger bulk of changes was done back in 2015 I frehed up
the stuff for the current Debian version 10.
A thing is that no longer a maintainance script is used to clean the
(global) cache directory but systemd-tmpfile and another thing is that
the RFCPATHS search path for (local) RFC documents is now also globally
set so that user login scripts are able to extend the search path (might
be changed to usage of systemd-environment.d in the future).